requirements.txt
Flask
psycopg2-binary==2.9.9
pandas==2.1.1
python-dotenv==1.0.0
XlsxWriter==3.1.2
